Clean command fully works

+++ src/Command/CommandClean.cpp
@@ -7,17 +7,18 @@ namespace Command{
     const std::vector<std::filesystem::path> files_to_delete = {
       pro->project_path / "CMakeCache.txt",
       pro->project_path / "install_manifest.txt",
+      pro->project_path / "cmake_install.cmake",
       pro->project_path / "Makefile"
     };
     const std::vector<std::filesystem::path> dirs_to_delete = {
-      pro->project_path / "CMakeFiles",
-      pro->project_path / "_deps"
+      pro->project_path / "CMakeFiles/",
+      pro->project_path / "_deps/"
     };
 
     for(auto& file: files_to_delete){
       if(std::filesystem::exists(file)){
-        std::cout << "Deleting: " << file << std::endl;
         try{
+          std::cout << "Deleting: " << file << std::endl;
           std::filesystem::remove(file);
         } catch(std::exception& e){
           std::cout << e.what() << std::endl;
@@ -27,8 +28,8 @@ namespace Command{
     }
     for(auto& dir: dirs_to_delete){
       if(std::filesystem::exists(dir)){
-        std::cout << "Deleting: " << dir << std::endl;
         try{
+          std::cout << "Deleting: " << dir << std::endl;
           std::filesystem::remove_all(dir);
         } catch(std::exception& e){
           std::cout << e.what() << std::endl;
@@ -58,14 +59,17 @@ namespace Command{
 
     //TODO: delete cache files and deps
     std::cout << "Cleaning: " << std::endl;
-    std::cout << pro->project_path / "build/*" << std::endl;
-    for(auto& p: std::filesystem::directory_iterator(pro->project_path / "build")){
-      std::cout << p.path() << std::endl;
-      if(p.is_directory()){
-        std::filesystem::remove_all(p.path());
-      } else {
-        std::filesystem::remove(p.path());
+    if(std::filesystem::exists(pro->project_path / "build")){
+      for(auto& p: std::filesystem::directory_iterator(pro->project_path / "build")){
+        std::cout << "Deleting" << p.path() << std::endl;
+        if(p.is_directory()){
+          std::filesystem::remove_all(p.path());
+        } else {
+          std::filesystem::remove(p.path());
+        }
       }
+    }else{
+      std::cout << "Could not find ./build" << std::endl;
     }
 
     return true;
